+++ projects/ng2-gauge/src/lib/gauge.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { GaugeComponent } from './gauge.component';
+
+@NgModule({
+  declarations: [GaugeComponent],
+  imports: [CommonModule],
+  exports: [GaugeComponent],
+})
+export class GaugeModule {}