+++ projects/ng2-gauge/src/lib/shared/validators.ts
@@ -1,38 +1,58 @@
-import { GaugeProps, Sector } from './ng2-gauge.interface';
+import { GaugeProps, Sector } from './interfaces';
 
-const showError = (text: string) => {
-  console.error(`GaugeComponent: ${text}`);
+const error = (text: string, throwErr?: boolean) => {
+  const msg = `GaugeComponent: ${text}`;
+
+  if (throwErr) {
+    throw new Error(msg);
+  }
+  console.error(msg);
 };
 
 export const validate = (props: GaugeProps) => {
   if (!props.max) {
-    showError('The maximal value is not set.');
+    error('Missing "max" input property (or zero)', true);
   }
 
-  if (!(0 <= props.start && props.start <= 359) || !(0 <= props.end && props.end <= 359)) {
-    showError('The end and start must be between 0 and 359 degrees.');
+  if (props.max < 0) {
+    error('"max" input property cannot be negative.', true);
   }
 
-  if (props.light && props.light > props.max) {
-    showError('The red light trigger value cannot be greater than the max value of the gauge.');
+  if (
+    !(0 <= props.arcStart && props.arcStart <= 359) ||
+    !(0 <= props.arcEnd && props.arcEnd <= 359)
+  ) {
+    error(
+      'The scale arc end and start must be between 0 and 359 degrees.',
+      true,
+    );
   }
 
-  if (props.factor && props.factor >= props.max) {
-    showError('The factor cannot be greater than or equal to the max value.');
+  if (props.activateRedLightAfter && props.activateRedLightAfter > props.max) {
+    error(
+      'The red light trigger value cannot be greater than the max value of the gauge.',
+    );
   }
+
+  // if (props.scaleFactor && props.scaleFactor >= props.max) {
+  //   showError('The factor cannot be greater than or equal to the max value.');
+  // }
 
   if (props.sectors) {
     props.sectors.forEach((s: Sector) => {
+      if (s.from < 0 || s.to < 0) {
+        error('The sector bounds cannot be negative.', true);
+      }
+
       if (s.from > props.max || s.to > props.max) {
-        showError('The sector bounds cannot be greater than the max value.');
+        error('The sector bounds cannot be greater than the max value.', true);
       }
 
       if (s.from >= s.to) {
-        showError('The lower bound of the sector cannot be greater than or equal to the upper one.');
-      }
-
-      if (!s.color) {
-        showError(`Sector[${s.from}, ${s.to}] color is empty.`);
+        error(
+          'The lower bound of the sector cannot be greater than or equal to the upper one.',
+          true,
+        );
       }
     });
   }